requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
-e file:../odp-core
    # via -r requirements.in
alembic==1.15.1
    # via -r requirements.in
anyio==4.8.0
    # via starlette
async-timeout==5.0.1
    # via redis
authlib==1.5.1
    # via odp
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
coverage==7.6.12
    # via -r requirements.in
cryptography==44.0.2
    # via authlib
exceptiongroup==1.2.2
    # via
    #   anyio
    #   pytest
factory-boy==3.3.3
    # via -r requirements.in
faker==36.2.2
    # via
    #   -r requirements.in
    #   factory-boy
fastapi==0.115.11
    # via -r requirements.in
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
iniconfig==2.0.0
    # via pytest
mako==1.3.9
    # via alembic
markupsafe==3.0.2
    # via mako
ory-hydra-client==1.11.8
    # via odp
packaging==24.2
    # via pytest
pluggy==1.5.0
    # via pytest
psycopg2==2.9.10
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic[dotenv]==1.10.21
    # via
    #   fastapi
    #   odp
pytest==8.3.5
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via ory-hydra-client
python-dotenv==1.0.1
    # via pydantic
redis==5.2.1
    # via odp
requests==2.32.3
    # via odp
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.38
    # via
    #   -r requirements.in
    #   alembic
    #   sqlalchemy-utils
sqlalchemy-utils==0.41.2
    # via -r requirements.in
starlette==0.46.0
    # via
    #   -r requirements.in
    #   fastapi
tomli==2.2.1
    # via pytest
typing-extensions==4.12.2
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   sqlalchemy
    #   uvicorn
tzdata==2025.1
    # via faker
urllib3==2.3.0
    # via
    #   ory-hydra-client
    #   requests
uvicorn==0.34.0
    # via -r requirements.in
